// frontend/src/types/teams.ts
export type Player = {
  name: string;
  number?: string;
  position: string;
  age?: number;
  nationality?: string;
  image?: string;
  team: string;
  appearances?: number;
  goals?: number;
};

export type Staff = {
  name: string;
  role: string;
  team: string;
  image?: string;
  bio?: string;
};